move cam mask to constants
