--- pgdrive/pg_config/cam_mask.py
@@ -1,11 +1,0 @@
-from panda3d.core import BitMask32
-
-
-class CamMask:
-    MainCam = BitMask32.bit(9)
-    Shadow = BitMask32.bit(10)
-    RgbCam = BitMask32.bit(11)
-    MiniMap = BitMask32.bit(12)
-    PARA_VIS = BitMask32.bit(13)
-    DepthCam = BitMask32.bit(14)
-    ScreenshotCam = BitMask32.bit(15)
