+++ pgdrive/scene_creator/blocks/block.py
@@ -7,8 +7,7 @@ from panda3d.bullet import BulletBoxShape, BulletRigidBodyNode
 from panda3d.core import Vec3, LQuaternionf, BitMask32, Vec4, CardMaker, TextureStage, RigidBodyCombiner, \
     TransparencyAttrib, SamplerState, NodePath
 
-from pgdrive.constants import Decoration, BodyName
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import Decoration, BodyName, CamMask
 from pgdrive.scene_creator.blocks.constants import BlockDefault
 from pgdrive.scene_creator.lane.abs_lane import AbstractLane, LineType, LaneNode
 from pgdrive.scene_creator.lane.circular_lane import CircularLane
