+++ pgdrive/scene_creator/vehicle_module/lidar.py
@@ -5,8 +5,7 @@ from typing import Set
 import numpy as np
 from panda3d.bullet import BulletGhostNode, BulletSphereShape
 from panda3d.core import BitMask32, NodePath
-from pgdrive.constants import BodyName
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import BodyName, CamMask
 from pgdrive.scene_creator.highway_vehicle.behavior import IDMVehicle
 from pgdrive.scene_creator.vehicle.traffic_vehicle import PGTrafficVehicle
 from pgdrive.utils.asset_loader import AssetLoader
