+++ pgdrive/scene_creator/vehicle_module/mini_map.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
 from panda3d.core import Vec3, NodePath
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import CamMask
 from pgdrive.world.image_buffer import ImageBuffer
