+++ pgdrive/scene_creator/vehicle_module/rgb_camera.py
@@ -1,6 +1,6 @@
 from panda3d.core import Vec3, NodePath
 
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import CamMask
 from pgdrive.world.image_buffer import ImageBuffer
 from pgdrive.world.pg_world import PGWorld
