+++ pgdrive/scene_creator/vehicle_module/routing_localization.py
@@ -2,8 +2,7 @@ import logging
 
 import numpy as np
 from panda3d.core import BitMask32, LQuaternionf, TransparencyAttrib
-from pgdrive.constants import COLLISION_INFO_COLOR, RENDER_MODE_ONSCREEN
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import COLLISION_INFO_COLOR, RENDER_MODE_ONSCREEN, CamMask
 from pgdrive.pg_config.parameter_space import BlockParameterSpace, Parameter
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
 from pgdrive.scene_creator.lane.circular_lane import CircularLane
