+++ pgdrive/scene_creator/vehicle_module/vehicle_panel.py
@@ -1,6 +1,6 @@
 from panda3d.core import NodePath, PGTop, TextNode, Vec3, CardMaker
 
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import CamMask
 from pgdrive.world.image_buffer import ImageBuffer
 from pgdrive.world.pg_world import PGWorld
