+++ pgdrive/world/screen_shot_camera.py
@@ -1,6 +1,6 @@
 from panda3d.core import Vec3
 
-from pgdrive.pg_config.cam_mask import CamMask
+from pgdrive.constants import CamMask
 from pgdrive.scene_creator.vehicle_module import MiniMap
